% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/radiant.R
\docType{package}
\name{radiant.design}
\alias{radiant.design}
\alias{radiant.design-package}
\alias{radiant.design}
\title{radiant.design}
\usage{
radiant.design(run = "browser")
}
\arguments{
\item{run}{Run radiant.design in an external browser ("browser") or in the the Rstudio viewer ("viewer")}
}
\description{
radiant.design

Launch radiant.design in default browser or Rstudio Viewer
}
\details{
See \url{https://radiant-rstats.github.io/docs} for documentation and tutorials
}
