% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/radiant.R
\docType{package}
\name{radiant.example}
\alias{radiant.example}
\alias{radiant.example-package}
\title{radiant.example}
\usage{
radiant.example()
}
\description{
radiant.example

Launch Radiant in the default browser
}
\details{
See \url{http://vnijs.github.io/radiant} for documentation and tutorials
}

