% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doe.R
\name{summary.doe}
\alias{summary.doe}
\title{Summary method for doe function}
\usage{
\method{summary}{doe}(object, eff = TRUE, part = TRUE, full = TRUE, est = TRUE, dec = 3, ...)
}
\arguments{
\item{object}{Return value from \code{\link{doe}}}

\item{eff}{If TRUE print efficiency output}

\item{part}{If TRUE print partial factorial}

\item{full}{If TRUE print full factorial}

\item{est}{If TRUE print number of effects that will be estimable using the partial factorial design}

\item{dec}{Number of decimals to show}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summary method for doe function
}
\details{
See \url{https://radiant-rstats.github.io/docs/design/doe.html} for an example in Radiant
}
\examples{
c("price; $10; $13; $16", "food; popcorn; gourmet; no food") \%>\%
  doe() \%>\%
  summary()

}
\seealso{
\code{\link{doe}} to calculate results
}
