% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizer.R
\name{summary.randomizer}
\alias{summary.randomizer}
\title{Summary method for the randomizer function}
\usage{
\method{summary}{randomizer}(object, dec = 3, ...)
}
\arguments{
\item{object}{Return value from \code{\link{randomizer}}}

\item{dec}{Number of decimals to show}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the randomizer function
}
\details{
See \url{https://radiant-rstats.github.io/docs/design/randomizer.html} for an example in Radiant
}
\examples{
randomizer(rndnames, "Names", conditions = c("test", "control")) \%>\% summary()

}
\seealso{
\code{\link{randomizer}} to generate the results
}
