% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{summary.sample_size}
\alias{summary.sample_size}
\title{Summary method for the sample_size function}
\usage{
\method{summary}{sample_size}(object, ...)
}
\arguments{
\item{object}{Return value from \code{\link{sample_size}}}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary method for the sample_size function
}
\details{
See \url{https://radiant-rstats.github.io/docs/design/sample_size.html} for an example in Radiant
}
\examples{
result <- sample_size(type = "mean", err_mean = 2, sd_mean = 10)
summary(result)

}
\seealso{
\code{\link{sample_size}} to generate the results
}
