% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/radiant.R
\docType{package}
\name{radiant.design}
\alias{radiant.design}
\title{radiant.design}
\usage{
radiant.design(state, ...)
}
\arguments{
\item{state}{Path to state file to load}

\item{...}{additional arguments to pass to shiny::runApp (e.g, port = 8080)}
}
\description{
Launch radiant.design in the default web browser
}
\details{
See \url{https://radiant-rstats.github.io/docs/} for documentation and tutorials
}
\examples{
\dontrun{
radiant.design()
}
}
