% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/radiant.R
\docType{package}
\name{radiant.design}
\alias{radiant.design}
\alias{radiant.design-package}
\title{radiant.design}
\usage{
radiant.design()
}
\description{
radiant.design

Launch radiant.design in the default browser
}
\details{
See \url{https://radiant-rstats.github.io/docs} for documentation and tutorials
}
\examples{
\dontrun{
radiant.design()
}
}
