% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant.R
\name{radiant.design_viewer}
\alias{radiant.design_viewer}
\title{Launch radiant.design in the Rstudio viewer}
\usage{
radiant.design_viewer(state, ...)
}
\arguments{
\item{state}{Path to state file to load}

\item{...}{additional arguments to pass to shiny::runApp (e.g, port = 8080)}
}
\description{
Launch radiant.design in the Rstudio viewer
}
\details{
See \url{https://radiant-rstats.github.io/docs/} for documentation and tutorials
}
\examples{
\dontrun{
radiant.design_viewer()
}
}
